Update process.ca.sh
I found the serial output script to by unstable and was unable to fix the problem directly. Instead I check the filesize from the Process script and restart the serial output script if it is dead.

Note. This is for the latest release of Raspbian Jessie Lite assuming you created a service for each script in order to launch them at startup.