python_serial.service
This file must be added to /lib/systemd/system/ and enabled through systemctl for Jessie systems.

I found, on jessie, that the best way to launch on startup is to create a service. The above will directly launch the python_serial.py script.
This should also replace the Launch script in jessie vs Wheezy.